move spec coverage evaluation code to augen
