+++ augen/.gitignore
@@ -6,6 +6,8 @@ public/server-api.json
 **/*.js
 !**/ReqResCache.*
 !**/closestSpec.*
+!**/*Relevant*.*
+!**/evalSpec*.*
 !src/test/*.js
 !src/test/toyApp/*.js
 src/test/toyApp/checkers*/*
